Reduce precision of dummy matrix
The matrix only contains 0's and 1's, so it doesn't need to be int64. I
switched it to float32, since we're using multiplying other float32s.

+++ openghg_inversions/array_ops.py
@@ -58,7 +58,7 @@ def get_xr_dummies(
         stack_dim = "".join([str(dim) for dim in da.dims])
         da = da.stack({stack_dim: da.dims})
 
-    dummies = pd.get_dummies(da.values, dtype=int, sparse=return_sparse)
+    dummies = pd.get_dummies(da.values, dtype="float32", sparse=return_sparse)
 
     # put dummies into DataArray with the right coords and dims
     values = COO.from_scipy_sparse(dummies.sparse.to_coo()) if return_sparse else dummies.values
